# Implement killer moves heuristic