Performance comparison:

Before (https://github.com/jamesqo/chess-bot/commit/550aa035974aad268d7fd20ff75550e7b0b53882), for `1. e4 2. Nf3`:
```
(mtdf depth=7)
> searchtimes
e7e5 - 33947.6869ms
d8f6 - 45440.8911ms

(ids depth=7)
> searchtimes
e7e5 - 70732.981ms
d8f6 - 72649.6108ms
```

After (dabec81):
```
(mtdf depth=7)
> searchtimes
e7e5 - 13846.6281ms
d8f6 - 14579.3078ms

(ids depth=7)
> searchtimes
e7e5 - 35554.5377ms
d8f6 - 32610.7481ms
```

(we should figure out why IDS is performing slower than MTD-f)